Separate unit and integration tests for clear distinction
Moved tests with external dependencies to tests/integration/ to ensure unit tests
run fast without external dependencies:

- Moved TestRequestExceptionIntegration from test_github_monitor.py to
  tests/integration/test_github_monitor_integration.py (uses subprocess calls)
- Moved test_main_loop_error_handling.py to
  tests/integration/test_main_loop_error_handling.py (uses subprocess calls)

Unit tests (make test-unit):
- Run 96 tests in ~6-7 seconds
- No external dependencies (no subprocess, no real network calls)
- All mocked and use temporary files/directories
- Fast feedback for development

Integration tests (make test-integration):
- Include tests that require subprocess execution
- Include tests that may need external services
- Run separately to avoid slowing down unit test feedback

This ensures developers can run 'make test-unit' for fast feedback during
development without waiting for slower integration tests.
